Adds POM rewriting capabilities.
Also changes the transformation to work over a set of files instead of a
single one.

Since we need to get all the POM files before we even start transforming
any libraries (to know that there is a library that is using older
version of support library that is not supported). The transformation
process firstly scans POM files and then does transformations on the
rest of the files. After that it finally also modifies the POM files.

I still have a few TODOs but didn't want to have the CL even bigger.
- Fix reference to any dependency we have rewritten in all the
discovered POM files
- Hook the gradle plugin to work over a set of libraries

UPDATE: Pom files can define values using properties. So I have extended
the tool to support that. I have also added better version checking
(support for intervals) and all PomRewriteRules versions are now validated
before the process is started

Bug: b/64833662
Test: Added in PomDocumentTest
Change-Id: I4e270de6c4590259300772102781b81c6d9237d4