Fix ACTION_PROCESS_TEXT menu on selection in O
Android 8.0 (and 8.1) developers reported that the menu created on text
selection for ACTION_PROCESS_TEXT is not working properly. Specifically,
because of a bug, a number of options that should have been in the menu
were missing. The bug was fixed in platform (b/66939160).  This CL adds
in the support library a possibility for developers to fix the bug for
the affected devices.

Bug: 68991560
Test: ./gradlew support-compat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.v4.widget.TextViewCompatTest\#testSetCustomSelectionActionModeCallback_fixesBugInSdk26
Change-Id: Ib2348aef79a5153a40ed3ff39712879316cecab7