Fix scroll bar logic.
Before, the offset and size of the scroll bar was not being calculated
correctly. This lead to the behavior that when the list scrolled to the
end, it would jump all the way there. Now, it is correctly scaled.

Also, remove the minimum and maximum sizes for the scroll bar thumb. UX
has dictated that no min or max size is needed.

Test: build on same application and test with varying item sizes.
Change-Id: If486b1547dd2e466302dc6bca009a80cdf35c7b1
Fixes: 70812326