Implement option to specify an app header that disppears when scrolled.
The app header is implemented as an AppBarLayout within a
CoordinatorLayout. However, since the contents of the Drawer is
technically part of the Coordinator Layout, it makes it so that the
header also disappears if the drawer is scrolled.

To fix this, need to manually disable the scrolling when the drawer is
opened. Also, remove the AppBarLayout's background so that the drawer's
color shows through.

Test: Test on sample app
Bug: 69866567
Change-Id: I730b17c0034857c5a603751a1411c3b6b567ff6f