Fix crash in AppCompatEditText#getText
Because this method can potentially be called before the TextView
constructor runs, we can't call setText safely, as it will cause
other un-initialized variables to fail. In P we can do a best-effort
to return an editable version of the given text, but pre-P we can't
distinguish between the text still being un-initialized and set to
a non-editable value to we will just return null for both.

Note this only happens when you call getText before the constructor
for TextView can finish running, so should be a very rare case.

Bug: 71638589
Test: AppCompatEditTextTest
Change-Id: I5c6139cecde9ff50e7f81f340b9c5623189b38eb