Stop DialogFragment before dismissing the dialog
Instead of dismissing the Dialog immediately and
queuing up the FragmentTransaction to pop the
DialogFragment off the stack, rely on the
onDestroyView callback on the DialogFragment to
dismiss the Dialog.

This ensures that the Dialog is non-null when the
DialogFragment gets a callback to onStop(), ensuring
that lifecycle observers always have access to the
Dialog.

Test: DialogFragmentTest
BUG: 64400546
Change-Id: Ic4847fd9efc475df08c03feb4e1983a2c19a8ea8