AppBarLayout with RecyclerView fling sample app
When an AppBarLayout is used together with a RecyclerView, the
coordinates of the MotionEvent that are received by the RecyclerView
during a scroll down or a fling event would have inconsistent origins.
Therefore, if these events are passed to VelocityTracker to calculate
the fling velocity, they may cause unexpected behaviour, typically
resulting in higher than desired velocity.

A framework change has been made in P to change the behaviour of
VelocityTracker to only use the raw coordinates of MotionEvent when
calculating the velocity.

VelocityTracker can use different strategies for estimating the
velocity. In P and below, the "lsq2" strategy is used. This strategy was
not sensitive to the changes in the origin of the MotionEvents. However,
a new strategy under consideration, "impulse", is sensitive to these
changes.

The app being added is a repro case for the original bug. The fling
behaviour can be tested with different VelocityTracker strageties. The
strategy can be set with "setprop debug.velocitytracker.strategy
<name of strategy>".

Bug: 72263561
Test: 1) Run "setprop debug.velocitytracker.strategy impulse" in order
to switch to impulse algorithm from the default (currently) lsq2
2) Run the support-v7-demos with the provided AppBarLayout activity that
reproduces the original bug. Fling back and forth on the RecyclerView
inside the activity
3) Observe abnormal behavior when the app bar collapses or expands (on P
and below, with "impulse" strategy).

Change-Id: I692cfeaf48434b992990a29c1dd7695db45ccef0