Rename Processor methods to start/stopWork.
The term "cancelled" is overloaded in our codebase.
Currently, it is used for both a user-initiated cancel
as well as a removal/stopping of a runnable/future.
This is a small change that renames the Processor
methods to startWork and stopWork, leaving the term
"cancel" to refer to user-initiated operations and
their corresponding database states.

Test: updated and ran tests.
Change-Id: If08c70902c2117dbc579678fac26f68ff02ffe50