Correctly determine rescheduling on bg job stop.
Right now, we incorrectly rely on Processor.stopWork's
result to determine if we should reschedule.  This
completely ignores a user-initiated cancel (that failed
to stop the runnable) was issued, resulting in the
service's onStopJob being invoked.

The solution here is to have the scheduler keep track
of user-initiated cancels (while that scheduler/process
is alive).  Since we implicitly know that the process
won't be brought down without signalling onStopJob, we
can then check the scheduler to determine if the job's
been cancelled (which means don't reschedule) or not
(go ahead and reschedule).

Also added some tests.  For the FirebaseJobService
tests, I had to copy over a bunch of test files
because workmanager-firebase's tests don't depend on
workmanager's tests.  We should fix this down the
road.

Bug: 72831451
Test: Updated and ran tests.
Change-Id: I2d52c528de0ae9410535d7efe0d7d3f3934c2069