Implement a ListDialog.
This dialog will display a fixed array of Strings. Under the hood, it
utilizes the TextListItem to display these items.

There is no affordance for titles or buttons and no support of radio
buttons at the time. Users can attach a listener to be notified of which
item was clicked. When an item is clicked, the dialog will dismiss
itself.

Test: Test and deploy on sample app
Fixes: 70894054
Change-Id: If99f6565ae5d9ab31b3b26ba72de86f7dc0d09aa