Support RawQuery in paging data source
This CL fixes a bug where paged list data source would not generate
proper code when a RawQuery is provided.

To overcome this, I've created a RoomSQLiteQuery helper method that
creates a query from a given support query.

I've also added getArgCount to the SupportSQLiteQuery API, which was
previously requested (b/67038952).

Bug: 67038952
Bug: 72600425
Test: DataSourceFactoryTest
Change-Id: I76183d6f02e9809bdbdad2d24159900497828b1b