Fix stale onLoadFinished callbacks on detach
markForRedelivery() resets the callback so that
the next time the LifecycleOwner gets started the
data is redelivered. However, removeObserver would
null out mObserver, causing LoaderInfo to keep
references to previous LoaderCallbacks and continue to
call them.

This manifests itself when a Fragment is detached and
reattached (such as in a FragmentPagerAdapter) and calls
initLoader() after onCreate() but before onStart().

Test: new LoaderInfoTest tests pass
BUG: 74182171
Change-Id: I3c82ae02d7bcc138bfec348e7e20be4017440ab5