End removal animation prior to the View becoming fully visible.
Bug: 74051124

Previously, the endViewTransition() call was being made after the
animation completed by a full vsync. This allowed the View to
be briefly visible after the view animation had completed.

This CL uses an OnPreDrawListener instead of posting the runnable,
allowing the animation to be ended prior to the next frame.

Test: manual
Test: ran fragment tests
Change-Id: Iaa74bfdb4bc79076fd0f31af3451202cc116c50a