Avoid improperly removing views during transitions
Bug: 65715616

There are two places where views were being removed improperly
during transitions. The first was when making a copy of a view.
Because of hardware bitmaps, views were being moved to the overlay
so that the image could be copied. This CL moves the view back
into its parent after copying the image.

The second location is where the view was being considered an
overlay. When a view is in the starting scene and the view in
the end scene cannot be used, the starting scene view was being
moved to the overlay. This is only valid when the view is alowed
to be removed. Instead, a bitmap copy of the view is moved to
the overlay.

Test: new FadeTest
Framework change: Ieb32b146cf65e3ff4ed6d7bb8325e74763dbd2d5
Change-Id: I99a7248bfd8a18ae922b4351bf79e2ff2e51a319
(cherry picked from commit 2e40b358face45adb6c6479037549750552079d0)