[WebView Support Library] Add initial WebViewClientCompat
This is the first CL in a series to implement WebViewClientCompat.

This adds the WebViewClientCompat class, implementing the boundary
interface. This provides real implementations for 3 of the methods
(category 1 APIs) and dummy implementations for the other 2 (category 2
APIs). All other APIs are inherited from WebViewClient, as those are
defined in API level 21 or earlier.

The dummy implementations have TODOs to resolve these in follow-up CLs,
after the initial plumbing has landed in chromium.

Definitions for category 1 and 2 can be found on the design doc:
http://go/wv-support-library-callbacks.

Bug: 73151460
Test: Manual - build a test app with WebViewClientCompat
Change-Id: I13087607bf46da0278db32793559da9eafb2e327