Fix constraints for 4 item single page nav drawer
Previously, the 4 item single page nav drawer's center row specified:
    app:layout_constraintBottom_toTopOf="parent"
    app:layout_constraintTop_toTopOf="parent"
The goal was to center these vertically in the parent, and at least in
some cases it worked, but it's somewhat non-sensical and seemed to be
off screen in other cases.

This change makes those constraints:
    app:layout_constraintBottom_toBottomOf="parent"
    app:layout_constraintTop_toTopOf="parent"

At the very least, it will make more sense. It should also clear up the
problem mentioned in the bug.

Test: Verified with sample app
Bug: 75986386
Change-Id: I542b398506239e8ac474141e0269bc4291798c3c