FIx RxJava Single & Maybe queries
We had a bug where the generated code for RxJava's Single and
Maybe would release the RoomSQLQuery object they use after the
first run. This would create an issue if the developer hold onto
the Single / Maybe and try to use it again. By that time, we would
re-use the query holder for another query.

Now generated code uses finalizer to release the query. It is
questionable whether it is worth doing the finalize block but
for consistency w/ other generated code, i'm keeping it.

Bug:76031240
Test: RxJava2Test
Change-Id: I6005995addfc8aa529c47e8f1d5bc653bc7f1415