Remove struct used to hold dialog data.
Instead, just access this data off the Builder directly. If the values
need to be held on to longer than the constructor, then the dialog
should create individual fields for this.

The reasoning is that this reduces abstraction. The DialogData objects
also allowed values to change after its creation. This works for
Builder, but should not be the case for the dialog itself.

Refactoring in this way ensures the expected behavior.

Fixes: 78349123
Test: test on demo app for CarAlertDialog and CarListDialog. Ensured
functionality is still the same with no crashes.

Change-Id: I9316e76ca8691506fdab6726388a56c274f87014