Fix #getMessageFromBundle from throwing NPE.
It's possible that an app using an old version of the compat library
doesn't encode a KEY_SENDER extra when writing MessagingStyle.Message
bundle. If a listener running the current compat version attempts to
 #getMessageFromBundle, it will throw an NPE due to the legacy null
case. The fix is simply to pass along the null to Message where the
encoded Person should, correctly, be null.

Test: ./gradlew core:connectedCheck
Fixes: 78484177
Change-Id: I8ec2302aa46069bd1fa26589fefcd27960a42803