[WebView Support Library] per-method feature detection for SafeBrowsingResponse
This adds per-method feature detection for SafeBrowsingResponseCompat.
This is implemented following the same pattern as http://ag/3957137.

SafeBrowsingResponseCompat is implemented by SafeBrowsingResponseImpl,
which delegates to a SafeBrowsingResponse (frameworks instance) when on
the appropriate API level, or a boundary interface otherwise. If neither
is available, it throws an exception--this can only happen if the app
omits feature-flag checks.

Bug: 77808798
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I51a8d8d2b07c7784c5bf988fb4f754fe088103fe