Relax conditions for jetifier xml rewrite.
Jetifier had very restricted set of XML tags to rewrite. This was
causing failures and the list had to be extened. Because this does not
scale well I have relaxed the selectors and jetifier now tries to
rewrite basically anything in quotes or in tags.

Another tiny change is that Jetifier will move files no matter if it is
rewriting support library or not. This is related to the new strategy
that we will rewrite files that are not originally from SL but are in
the same namespace and satisfy our rules. By this we won't break usages
of internal APIs.

Test: Added
Change-Id: I20041897eaa98899294092bd88a9371d811026d6
Fix: 79111795
Bug: 64833662