Modify behavior of the drawer opening.
Restructure the drawer so that it is more inline with the
implementations of the drawer on the phone. When the drawer opens, it
will slide on top of the toolbar.

Before, the toolbar was technically on top of the drawer and had logic
to style itself as appearing below the drawer when it opened.

This refactor will allow for better maintainability in the future.

Change-Id: I0597625fd8dca8d2a2024bc98e6eca22edf4854f
Fixes: 80193724
Test: Test drawer opening behavior on demo app.