Add enqueueUniquePeriodicWork(Sync).
Unlike OneTimeWorkRequests, this API does not provide
begin/then syntax because you can't chain periodic work.
Internally, the implementation is identical to one-time
work, but since ExistingWorkPolicy.APPEND does not make
sense for PeriodicWork, I created a new enum for it,
ExistingPeriodicWorkPolicy.

I'm not totally satisfied with having two different
enums that are similar, so we can explore the potential
ramifications in the future.

Change-Id: Ic43289c6654bd2c031a432692a52cc5fb25e1991
Fixes: 79600647
Test: Added and ran tests.