Revert public API change to Fragment.SavedState.CREATOR
While it is important that the underlying CREATOR
object is a ClassLoaderCreator, we should not change
the public API to use ClassLoaderCreator, but instead
still have the public API as Creator.

This works since the framework uses instanceof to
change its behavior based on whether it is a
ClassLoaderCreator, not the reflected field type.

Test: existing tests pass
BUG: 28678246
Change-Id: Ib80355cdd4e9f9756453de8d5ce6436039748bbc