Allow small slices to be displayed at less than our “default” height
If height set on SliceView / available height is <= 48dp, enforce height
of 48dp (unless it’s a slider which will be bigger) when displayed as
small / large template.

Contents of row will be centered in this space / with appropriate padding.

If you’re in large mode the height always needs to at least show your
1st row.

Fixes: 80436569
Test: ./gradlew slice-view:connectedCheck
      manual - use SliceBrowser and adjust height of view to be
               < 48dp; == 48dp; > 500dp; check various slices in
               small / large / shortcut and make sure they behave as
               expected, notably:
               - input range slice should always show text + slider row
               - small / large slices should never display < 48dp high

Change-Id: I94b133c082098ccc92887fb4664769de487bfcda