Ensure Fragment animations are removed even if they never end()
Bug: 109856757
Bug: 111317186

Fragment animations can stop before completion when a window
covers up the Activity. When this happens, the view remains
in the disappearing children list and wil then always be shown.

This CL forces the call that removes the animation from the view
when the animation stops playing, and the view is thus removed
from the disappearing children list.

I could not find a way to test this CL without using reflection
and P negates the ability to use reflection. So, no automated
tests.

Test: I created a manual test that reproduced 100% of the time
Test: Tested application that exhibited the error
Change-Id: I89d4abae4b4c5f914fe88c07a67614fedfb645a3