Fixed RecyclerViewLayoutTest#smoothScrollWithRemovedItemsAndRemoveItem
The test was seemingly relying on querying for a value at a
nondeterministic time within a nondeterministic timeframe.  Now the
test keeps track of all values set in a colleciton, and then at the
end makes sure that the value that was being queried for simply
exists in the collection.

Bug: 33659102
Fixes: 33659102
Test: RecyclerViewLayoutTest#smoothScrollWithRemovedItemsAndRemoveItem

Change-Id: I3ebaac29a540052902ef9f448c9584696cbba8fa