Replace OpenHelper's wrapped DB reference when the received DB from super mismatches.
If an exception is thrown during any of the OpenHelper methods the
SQLiteOpenHelper recognizes a failed DB initialization and returns a
closed database object. Upon using the DB again the helper will re-try
to open the DB causing Room's wrapped DB reference to be out of sync.
This change detects when the wrapped DB Room is delegating to has
changed so that he can update its reference.

Bug: 111504749
Test: ./gradlew :room:integration-tests:testapp:cC
Change-Id: I2d5f7d91a64a71e886609e2c4625382cc4c3e5ac