Fix back stack inconsistency when detecting back pressed.
FragmentNavigator's back stack was becoming inconsistent with the
FragmentManager's back stack due to in-between multiple pending
operations callbacks occurring that would cause the difference between
the new count and the final count to be equal and thus causing the
amount of pending operations not to be decreased.

This changes takes a slightly different strategy, instead of counting
pending operations, we just flag that we have pending operations and we
don't consider them done until the FragmentManager's back stack is equal
to the Navigator's back stack. Consuming back stack callbacks as long
as there are pending operations so that we don't accidentally detect a
back press. Since a back press causes a popBackStackImmediate() we have
confidence pending operations will be executed causing the Navigator to
reset the flag to later then detect the back press pop.

Bug: 111884848
Bug: 111907708
Test: FragmentNavigatorTest, additionally, manually verified fix works
      on both sample projects provided.

Change-Id: I51be4d6a974ecbeeaa1f63d764edabb230ae85e2