Make ExoPlayer wrapper thread safe
ExoPlayer itself is not thread-safe, so this change synchronizes access
to it and passing the internal task looper when creating the player to
ensure that callbacks occur on the task thread.

Move remaining implemented player methods over to use the command queue
and implement get/setPlayerVolume.

Also add a supression for lint RestrictedApi checks which don't
currently work correctly in Android Studio due to not setting the
groupId for the sources.

Test: set DEBUG_USE_EXOPLAYER to true and played a video with VideoView2
Bug: 80232248
Change-Id: I0695b7753b1a08f84fce06ffec4bfb3058655e06